requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.8.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.2
idna==3.10
imageio==2.35.1
ipdb==0.13.13
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
sentry-sdk==2.16.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.3
wcwidth==0.2.13
Werkzeug==3.0.4
